refactored the tiny server framework code to handle POSIX and Windows systems in different files
